Make script runnable from terminal
Currently you need to pass the script to the python interperter to try the
game outside of an interpreter. By chmod +x and adding a shebang
you can just run the game from the terminal.

Note: You need to be in the main folder to load the game.
It doesn't work if run within game/